+++ soh/soh/Enhancements/randomizer/randomizer_grotto.h
@@ -35,7 +35,7 @@ void Grotto_ForceGrottoReturnOnSpecialEntrance(void);
 void Grotto_ForceGrottoReturn(void);
 void Grotto_ForceRegularVoidOut(void);
 void Grotto_SanitizeEntranceType(void);
-s16 Grotto_GetRenamedGrottoIndexFromOriginal(s8 content, s8 scene);
+s16 Grotto_GetRenamedGrottoIndexFromOriginal(s8 content, s16 scene);
 s8 Grotto_CurrentGrotto();
 #ifdef __cplusplus
 };
